move to `failover` package instead of `ha` and mark beta
