+++ policy/src/test/java/org/apache/brooklyn/policy/failover/ElectPrimaryTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.brooklyn.policy.ha;
+package org.apache.brooklyn.policy.failover;
 
 class ElectPrimaryTest {
