+++ policy/src/test/resources/org/apache/brooklyn/policy/failover/elect-primary-propagate-test.yaml
@@ -17,7 +17,7 @@
 # under the License.
 #
 brooklyn.policies:
-- type: org.apache.brooklyn.policy.ha.ElectPrimaryPolicy
+- type: org.apache.brooklyn.policy.failover.ElectPrimaryPolicy
   brooklyn.config:
     primary.stopped.wait.timeout: 0
     propagate.primary.sensors: [ sens1, sens2 ]
